# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blue Mogul Enterprise Bookkeeping AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, transfer]
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: reconciled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /transactions/{id}/reconcile:
    post:
      operationId: reconcileTransaction
      tags: [reconciliation]
      summary: Mark a transaction as reconciled
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcileTransactionBody"
      responses:
        "200":
          description: Reconciled transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        reconciled:
          type: boolean
        reconciledAt:
          type: string
          format: date-time
          nullable: true
        source:
          type: string
          enum: [manual, wave, stripe]
        sourceId:
          type: string
          nullable: true
        taxDeductible:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - description
        - amount
        - type
        - category
        - reconciled
        - source
        - taxDeductible
        - createdAt
    CreateTransactionBody:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        category:
          type: string
        subcategory:
          type: string
        taxDeductible:
          type: boolean
        notes:
          type: string
      required:
        - date
        - description
        - amount
        - type
        - category
    UpdateTransactionBody:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense, transfer]
        category:
          type: string
        subcategory:
          type: string
        taxDeductible:
          type: boolean
        notes:
          type: string
    ReconcileTransactionBody:
      type: object
      properties:
        bankAmount:
          type: number
        notes:
          type: string
      required:
        - bankAmount
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error